% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_attribution}
\alias{add_attribution}
\title{Add Attribution Caption to Plots}
\usage{
add_attribution()
}
\value{
A ggplot2 \code{labs} object with a caption attribute if acknowledgments
are enabled; otherwise, \code{NULL}.
}
\description{
This function checks if acknowledgments are enabled in the
environment and, if so, adds an attribution caption to the plot using ggplot2.
If attributions are not enabled, it returns \code{NULL}.
}
