% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_query_term.R
\name{add_new_query_term}
\alias{add_new_query_term}
\title{Add a new query term from the output of collect_query_terms that is split into a list}
\usage{
add_new_query_term(list, list_index_number, ...)
}
\arguments{
\item{list}{list object from the collect_query_terms function}

\item{list_index_number}{the index number of list that the query is added to}

\item{...}{character strings of the additional queries that will be added to the list}
}
\value{
a list object with the query/queries added at the appropriate index
}
\description{
Add a new query term from the output of collect_query_terms that is split into a list
}
