% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjust_display}
\alias{adjust_display}
\title{Adjust display}
\usage{
adjust_display(osm_object)
}
\arguments{
\item{osm_object}{The osm_object with stored bbox}
}
\description{
This function adjust displays and should be called last after adding geoms
}
