% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adjust_viewport}
\alias{adjust_viewport}
\title{Adjust the viewport for map visualization}
\usage{
adjust_viewport(osm_object)
}
\arguments{
\item{osm_object}{An OSM object containing a bbox attribute.}
}
\value{
A ggplot2 coordinate (coord_sf) object with adjusted x and y limits.
}
\description{
This function adjusts the viewport for map visualization using ggplot2.
It modifies the x and y limits based on the bounding box (bbox) of the OpenStreetMap (OSM) object.
}
\details{
The function calculates new x and y limits by adding and subtracting a twentieth of the bbox's width and height
from the respective minimum and maximum x and y values. This removes the margin around the map, enhancing visibility.
}
