% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_staged_to_key.R
\name{append_staged_to_key}
\alias{append_staged_to_key}
\title{Adds PROGRESS_LIST$Staged to KEY}
\usage{
append_staged_to_key(progress_list, path_to_key_csv)
}
\description{
Adds PROGRESS_LIST$Staged to KEY
}
