% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{berlinaccidents}
\alias{berlinaccidents}
\title{Berlin accidents dataset for 2021}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 11267 rows and 23 columns.
}
\usage{
berlinaccidents
}
\description{
#' This dataset was produced from the Amt für Statistik Berlin-Brandenburg data available from
\url{https://www.govdata.de/web/guest/daten/-/details/strassenverkehrsunfalle-nach-unfallort-in-berlin-2021}.
}
\keyword{datasets}
