% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_bucket_denovos.R
\name{bulk_bucket_denovos}
\alias{bulk_bucket_denovos}
\title{Bulk flags rows for denovo cui based on row numbers}
\usage{
bulk_bucket_denovos(progress_list, denovo_row_numbers, bucket)
}
\arguments{
\item{denovo_row_numbers}{denovo_row_numbers of progress_list$PARKED_FOR_DENOVO}

\item{bucket}{bucket name for downstream workflow: radiology, experimental drugs, precoordinated (combination of multiple existing concepts), qualifier}
}
\description{
Bulk flags rows for denovo cui based on row numbers
}
