% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_edit_external_followup.R
\name{bulk_edit_external_followup}
\alias{bulk_edit_external_followup}
\title{Bulk flag rows for external followup}
\usage{
bulk_edit_external_followup(progress_list, row_numbers,
  followup_request_edit = NULL, followup_question_edit = NULL)
}
\arguments{
\item{row_numbers}{single or range of row_numbers in PROGRESS_LIST$PARKED_FOR_EXTERNAL_FOLLOWUP}
}
\description{
Bulk flag rows for external followup
}
