% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_flag_rows_for_external_followup.R
\name{bulk_flag_rows_for_external_followup}
\alias{bulk_flag_rows_for_external_followup}
\title{Bulk flag rows for external followup}
\usage{
bulk_flag_rows_for_external_followup(progress_list, row_numbers,
  bulk_followup_request = "", bulk_followup_question = "")
}
\arguments{
\item{row_numbers}{single or range of row_numbers in PROGRESS_LIST$QUEUE}
}
\description{
Bulk flag rows for external followup
}
