% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/bulk_timestamp_rows_for_completed_external_followup.R
\name{bulk_timestamp_rows_for_completed_external_followup}
\alias{bulk_timestamp_rows_for_completed_external_followup}
\title{Bulk changes PROGRESS_LIST$PARKED_FOR_EXTERNAL_FOLLOWUP and stages those changes}
\usage{
bulk_timestamp_rows_for_completed_external_followup(progress_list,
  row_numbers)
}
\arguments{
\item{progress_list}{full PROGRESS_LIST object}

\item{row_numbers}{row_numbers of PROGRESS_LIST$PARKED_FOR_EXTERNAL_FOLLOWUP}
}
\description{
Bulk changes PROGRESS_LIST$PARKED_FOR_EXTERNAL_FOLLOWUP and stages those changes
}
