% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\docType{data}
\name{cartographr_env}
\alias{cartographr_env}
\title{Package configuration environment}
\format{
An object of class \code{environment} of length 4.
}
\usage{
cartographr_env
}
\description{
This environment stores configuration settings for the package.
It is initialized when the package is loaded and can be modified
using the \verb{set_*()} functions.
}
\keyword{internal}
