% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_query_terms.R
\name{collect_query_terms}
\alias{collect_query_terms}
\title{Interactively collect phrases that will be used in the WHERE STR clause in the SQL statement}
\usage{
collect_query_terms(input_dataframe, output_data_name, script_step_number)
}
\arguments{
\item{input_dataframe}{dataframe in the KEY format that contains the column of source concepts that need to be mapped}

\item{script_step_number}{the number assigned to the script this function is being run on. This is important because the R object will be saved with the script number as a prefix}

\item{output_dataframe_name}{character vector of length 1 that will be the name of the output data, which will be list split based on concept and additional column of the new terms}
}
\description{
Interactively collect phrases that will be used in the WHERE STR clause in the SQL statement
}
