% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_table_of_contents.R
\name{create_new_table_of_contents}
\alias{create_new_table_of_contents}
\title{Creates a new table of contents from arguments in format: {term}:{definition}}
\usage{
create_new_table_of_contents(basename, ...)
}
\arguments{
\item{basename}{unique basename that will be the output csv filename in the format CATALOGUE/TABLE_OF_CONTENTS_{basename}.csv as well as the column name}
}
\description{
Creates a new table of contents from arguments in format: {term}:{definition}
}
