% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_progress_list.R
\name{create_progress_list}
\alias{create_progress_list}
\title{Takes IDENTITY and KEY and creates a PROGRESS LIST based on the status of the concept}
\usage{
create_progress_list(identity, key)
}
\arguments{
\item{identity}{output from identity functions}

\item{key}{output from key functions}
}
\description{
Takes IDENTITY and KEY and creates a PROGRESS LIST based on the status of the concept
}
