% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crime}
\alias{crime}
\title{Manhattan crime dataset}
\format{
A data frame with columns representing various attributes of crimes such as type, location, date, and time.
}
\source{
New York City Police Department (NYPD) \url{https://catalog.data.gov/}
}
\usage{
crime
}
\description{
This dataset encompasses all reported felony, misdemeanor, and violation crimes as recorded by the New York City Police Department (NYPD) for the borough of Manhattan, starting from the year 2004.
The data is sourced from the public domain and is available for analysis and research purposes. It provides a comprehensive overview of crime patterns and can be utilized for developing crime prevention strategies, conducting sociological research, and enhancing public awareness.
}
\keyword{datasets}
