% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographr.R
\name{cutout}
\alias{cutout}
\title{Preprocessing of a circular map}
\usage{
cutout(osm, boundary = "rect")
}
\arguments{
\item{osm}{The osm_object with stored bbox}
}
\value{
A preprocessed osm object
}
\description{
This function uses sensible defaults to remove unnecessary data from the osm object
}
