% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make_border_ring}
\alias{.make_border_ring}
\title{Build a border ring around crop_extent}
\usage{
.make_border_ring(
  crop_extent,
  lat = NULL,
  lon = NULL,
  border_width = 0.001,
  width_unit = c("rel", "m", "mm")
)
}
\description{
Build a border ring around crop_extent
}
\keyword{internal}
