% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographr.R
\name{.plot_map}
\alias{.plot_map}
\title{Plot a map}
\usage{
.plot_map(osm, palette = "imhof")
}
\arguments{
\item{osm}{OSM object to plot}

\item{palette}{Color theme applied to the plot}

\item{circle}{If TRUE draw circle plot}
}
\description{
This function plots a map
}
\keyword{internal}
