% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_query_mrconso.R
\name{double_query_mrconso}
\alias{double_query_mrconso}
\title{Searching mrconso with 2 LIKE phrases}
\usage{
double_query_mrconso(search_word_01, search_word_02)
}
\description{
Searching mrconso with 2 LIKE phrases
}
