% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_flag_rows_for_denovo_cui.R
\name{flag_rows_for_denovo_cui}
\alias{flag_rows_for_denovo_cui}
\title{Bulk flags rows for denovo cui based on row numbers}
\usage{
flag_rows_for_denovo_cui(progress_list, row_numbers,
  bulk_denovo_cui_justification = "")
}
\description{
Bulk flags rows for denovo cui based on row numbers
}
