% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_corpus.R
\name{generate_corpus}
\alias{generate_corpus}
\title{Generate the final corpus of words for the data}
\usage{
generate_corpus(valueset_output, token_output)
}
\arguments{
\item{valueset_output}{output from get_valueset_function}

\item{token_output}{output from get_tokens function}
}
\description{
Generate the final corpus of words for the data
}
