% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_next_id_number.R
\name{generate_next_id_number}
\alias{generate_next_id_number}
\title{Generate next ID number}
\usage{
generate_next_id_number(id_vector)
}
\arguments{
\item{id_vector}{vector of id numbers that is already existing. NA values are ignored.}
}
\value{
character vector of length 1 of the next ID number by random addition of a value between 1 and 9
}
\description{
Generate next ID number
}
