% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographr.R
\name{get_border}
\alias{get_border}
\title{Calculate rectengular border given the coordinates}
\usage{
get_border(lat, lon, offlat, offlon)
}
\arguments{
\item{lat}{Latitude WSG84}

\item{lon}{Latitude WSG84}

\item{offlat}{offset Latitude}

\item{offlon}{offset Longitude}
}
\value{
The border of the map in Coordinates
}
\description{
This function calculates correct borders
}
