% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_circle}
\alias{get_circle}
\title{Generate bounding circle}
\usage{
get_circle(lat, lon, y_distance, x_distance)
}
\arguments{
\item{lat}{Latitude WGS84}

\item{lon}{Longitude WGS84}

\item{y_distance}{Y distance in meters}

\item{x_distance}{X distance in meters}
}
\value{
The circle
}
\description{
This function generates a bounding circle
}
