% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corpus_from_tokens_list.R
\name{get_corpus_from_tokens_list}
\alias{get_corpus_from_tokens_list}
\title{Creates a corpus from the output of get_tokens_from_valueset()
Corpus: all the unique tokens belonging to the field in question}
\usage{
get_corpus_from_tokens_list(tokens_list)
}
\arguments{
\item{tokens_list}{output from get_tokens_from_valueset}
}
\description{
Creates a corpus from the output of get_tokens_from_valueset()
Corpus: all the unique tokens belonging to the field in question
}
