% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_possible_cui.R
\name{get_first_possible_cui}
\alias{get_first_possible_cui}
\title{Gets first possible cui for the input phrase in English and ISPREF = 'Y'}
\usage{
get_first_possible_cui(search_word)
}
\description{
Gets first possible cui for the input phrase in English and ISPREF = 'Y'
}
