% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_hexagon}
\alias{get_hexagon}
\title{Generate bounding hexagon}
\usage{
get_hexagon(lat, lon, y_distance, x_distance)
}
\arguments{
\item{lat}{Latitude WGS84}

\item{lon}{Longitude WGS84}

\item{y_distance}{Y distance in meters}

\item{x_distance}{X distance in meters}
}
\value{
The hexagon
}
\description{
This function generates a bounding hexagon
}
