% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{get_palette}
\alias{get_palette}
\title{Create a color theme for maps}
\usage{
get_palette(palette)
}
\arguments{
\item{palette}{The color palette to use.
Can be one of "macau", "barcelona", "night", "arkham", "imhof",
"bw", "bwinv", "amsterdam", "berlin", "wesanderson",
"modern", "terraqua", "bejing", "darkmode", or a named list for a custom palette.
If a named list is provided, it should contain color hex codes for each map element.
If \code{NULL} or an unrecognized name is provided, the function will throw an error.}
}
\value{
A list containing color settings for the map elements.
}
\description{
This function creates a color theme to be used with \code{plot_map()}. It can accept a predefined palette name or a custom palette provided as a named list.
}
\examples{
# Standard usage with a character argument for a predefined palette
plot_map(osm, palette = 'tropical')

# Custom palette creation using a named list for a simple black and white palette
custom_palette <- list(
  palette_building = c("#000000", "#FFFFFF"),
  water = "#000000",
  green = "#FFFFFF",
  beach = "#000000",
  parking = "#FFFFFF",
  street = "#000000",
  background = "#FFFFFF",
  railway = "#000000",
  hatched = TRUE
)
plot_map(osm, palette = custom_palette)
}
