% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographr.R
\name{get_scaling}
\alias{get_scaling}
\title{Get scaling for a paper format}
\usage{
get_scaling(format)
}
\arguments{
\item{format}{Paper format}
}
\value{
The scaling.
}
\description{
Retrieve correct scaling value based on size A1 = default
}
