% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcards_utils.R
\name{get_theme}
\alias{get_theme}
\title{Create color theme}
\usage{
get_theme(palette)
}
\arguments{
\item{palette}{The color palette}
}
\value{
A color palette for plot_map()
}
\description{
This function creates a theme
}
