% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tokens_for_target_vars.R
\name{get_tokens_for_target_vars}
\alias{get_tokens_for_target_vars}
\title{Get tokens for the valueset object outputed by get_valueset_for_target_vars}
\usage{
get_tokens_for_target_vars(valueset_output)
}
\arguments{
\item{valueset_output}{output from get_valueset_for_target_vars}
}
\description{
Get tokens for the valueset object outputed by get_valueset_for_target_vars
}
