% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tokens_from_valueset.R
\name{get_tokens_from_valueset}
\alias{get_tokens_from_valueset}
\title{Get list of tokens for a list of valuesets
Tokens: all unique lowercase, split at single spaces words}
\usage{
get_tokens_from_valueset(valueset_list)
}
\arguments{
\item{valueset_list}{output from get_valueset()}
}
\description{
Get list of tokens for a list of valuesets
Tokens: all unique lowercase, split at single spaces words
}
