% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valueset.R
\name{get_valueset}
\alias{get_valueset}
\title{Returns the unique values for a given value field in a list}
\usage{
get_valueset(dataframe, value_variable, major_id_variable, ...)
}
\arguments{
\item{value_variable}{column containing the value}

\item{major_id_variable}{grouping variable used to split the dataframe}

\item{...}{other id_variables to supplement the output (ie the actual labels to the id variable)}
}
\description{
Returns the unique values for a given value field in a list
}
