% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valueset_for_target_vars.R
\name{get_valuesets_for_target_variables}
\alias{get_valuesets_for_target_variables}
\title{Get a list of value sets per target variable}
\usage{
get_valuesets_for_target_variables(dataframe, id_var, target_vars)
}
\arguments{
\item{id_var}{id variable that will tag the output}

\item{target_vars}{character vector of 1 or greater of target variable names}
}
\description{
Get a list of value sets per target variable
}
