% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity_from_gs.R
\name{identity_from_gs}
\alias{identity_from_gs}
\title{Gets IDENTITY data and cleans up based on parameters}
\usage{
identity_from_gs(gsheet_name, gsheet_tab, target_cols,
  keep_all_cols = FALSE, dont_trimws = FALSE, remove_all = "\\t|\\n",
  include_duplicates = FALSE, log = TRUE)
}
\arguments{
\item{gsheet_name}{exact name of the Google Sheet that has the IDENTITY table. If there are more than 1 Google Sheets with this name, an error will be returned. An "IDENTITY_ID" column is required.}

\item{gsheet_tab}{name of the Google Sheet tab that has the IDENTITY table.}

\item{keep_all_cols}{TRUE if all the variables are wanted in the output. Otherwise, only the target_cols will be selected.}

\item{dont_trimws}{TRUE if white spaces on both the left and right side are not desired in the output}

\item{remove_all}{character string of length 1 representing the string of artifact and other values to be removed from the final output separated by the OR operator}

\item{include_duplicates}{TRUE if all the duplicates are desired in the output}
}
\description{
Gets IDENTITY data and cleans up based on parameters
}
