% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inner_join_key.R
\name{inner_join_key}
\alias{inner_join_key}
\title{Inner join data with a key on a single join column and a single column in the key that will sub in the dataframe}
\usage{
inner_join_key(dataframe, key, join_col, key_sub_col)
}
\arguments{
\item{dataframe}{dataframe to map to key}

\item{key}{dataframe with variables that needed to be joined into the dataframe}

\item{join_col}{column to join on. Must be present in botrh the dataframe and key}

\item{key_sub_col}{column to sub. Can be absent in the dataframe, but must be present in the key}
}
\description{
Inner join data with a key on a single join column and a single column in the key that will sub in the dataframe
}
