% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_double_term_search.R
\name{interactive_double_term_search}
\alias{interactive_double_term_search}
\title{Performs a double phrase search in MRCONSO rowwise in progress_list$QUEUE}
\usage{
interactive_double_term_search(progress_list)
}
\description{
Performs a double phrase search in MRCONSO rowwise in progress_list$QUEUE
}
