% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_single_term_search.R
\name{interactive_single_term_search}
\alias{interactive_single_term_search}
\title{Performs a single phrase search in MRCONSO rowwise in progress_list$QUEUE}
\usage{
interactive_single_term_search(progress_list)
}
\description{
Performs a single phrase search in MRCONSO rowwise in progress_list$QUEUE
}
