% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_from_gs.R
\name{key_from_gs}
\alias{key_from_gs}
\title{Gets KEY data and cleans up based on parameters}
\usage{
key_from_gs(gsheet_name, gsheet_tab, log = TRUE)
}
\arguments{
\item{gsheet_name}{exact name of the Google Sheet that has the KEY table. If there are more than 1 Google Sheets with this name, an error will be returned.}

\item{gsheet_tab}{name of the Google Sheet tab that has the KEY table. Requirements: KEY_TIMESTAMP, KEY_FIELD, and KEY_CONCEPT_NAME as the grouping variable.}
}
\description{
Gets KEY data and cleans up based on parameters
}
