% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_native_attributes.R
\name{lookup_native_attributes}
\alias{lookup_native_attributes}
\title{Looks up native attributes from IDENTITY}
\usage{
lookup_native_attributes(identity_id, key_field, key_concept_name)
}
\arguments{
\item{identity_id}{IDENTITY_ID}

\item{key_field}{name of target field}

\item{key_concept_name}{name of concept}
}
\description{
Looks up native attributes from IDENTITY
}
