% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_parked_for_denovo_to_bucket.R
\name{observe_parked_for_denovo_to_bucket}
\alias{observe_parked_for_denovo_to_bucket}
\title{Observe the PROGRES_LIST$PARKED_FOR_DENOVO to bucket these concepts}
\usage{
observe_parked_for_denovo_to_bucket(progress_list)
}
\description{
Observe the PROGRES_LIST$PARKED_FOR_DENOVO to bucket these concepts
}
