% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_parked_for_external_followup.R
\name{observe_parked_for_external_followup}
\alias{observe_parked_for_external_followup}
\title{Observe the PROGRES_LIST$PARKED_FOR_EXTERNAL_FOLLOWUP to map}
\usage{
observe_parked_for_external_followup(progress_list)
}
\description{
Observe the PROGRES_LIST$PARKED_FOR_EXTERNAL_FOLLOWUP to map
}
