% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{osm}
\alias{osm}
\title{OSM SoHo New York simple features}
\format{
An named list of objects of class \code{sf}
with rows and columns corresponding to the features and their attributes.
}
\source{
OpenStreetMap
}
\usage{
osm
}
\description{
This dataset contains the Simple Features (sf) representation for Soho, New York.
It includes various geographical and spatial attributes relevant to the area.
}
\examples{
data("osm")
}
\keyword{datasets}
