% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcards_utils.R
\name{preprocess_map}
\alias{preprocess_map}
\title{Preprocessing map}
\usage{
preprocess_map(osm, scaling = get_scaling("A4"), circle = F)
}
\arguments{
\item{osm}{The osm_object with stored bbox}

\item{scaling}{scaling applied to the plot}
}
\value{
A preprocessed osm object
}
\description{
This function uses sensible defaults to remove unnecessary data from the osm object
and sets the output size for printing.
}
