% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographr.R
\name{preprocess_map_circular}
\alias{preprocess_map_circular}
\title{Preprocessing of a circular map}
\usage{
preprocess_map_circular(osm)
}
\arguments{
\item{osm}{The osm_object with stored bbox}
}
\value{
A preprocessed osm object
}
\description{
This function uses sensible defaults to remove unnecessary data from the osm object
}
