% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographr.R
\name{print.cartographr_ggplot}
\alias{print.cartographr_ggplot}
\title{Print Custom ggplot Object}
\usage{
\method{print}{cartographr_ggplot}(x, ...)
}
\arguments{
\item{x}{A custom ggplot object with an attribute "osm" containing osm data.}

\item{...}{Additional arguments passed to the print method.}
}
\value{
The function does not return a value but prints the ggplot object.
}
\description{
This function takes a custom ggplot object, extracts the ggplot object,
retrieves the bounding box (bbox) from the associated osm data, and adjusts
the display before printing the plot.
}
