% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{print_config}
\alias{print_config}
\title{Print configuration details}
\usage{
print_config()
}
\value{
This function does not return any value; it only prints the configuration details.
}
\description{
This function prints the configuration details stored in the current environment.
It iterates through the variables in the environment and displays their names and values.
}
\examples{
print_config()

}
