% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_list.R
\name{progress_list}
\alias{progress_list}
\title{Fresh upload of IDENTITY and KEY to create a new PROGRESS_LIST object}
\usage{
progress_list(path_to_identity_csv, path_to_key_csv,
  filter_out_blank_concepts = TRUE)
}
\description{
Fresh upload of IDENTITY and KEY to create a new PROGRESS_LIST object
}
