% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_list_staged_to_key.R
\name{progress_list_staged_to_key}
\alias{progress_list_staged_to_key}
\title{Adds PROGRESS_LIST$Staged to KEY}
\usage{
progress_list_staged_to_key(progress_list, path_to_key_csv)
}
\description{
Adds PROGRESS_LIST$Staged to KEY
}
