% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_mrconso_net.R
\name{qa_mrconso_net_glossary}
\alias{qa_mrconso_net_glossary}
\title{This function will make sure that the GLOSSARY and list of files in NETS/ match.}
\usage{

  qa_mrconso_net_glossary(path_to_glossary = "/Users/meerapatel/GitHub/MSK_KMI_Enterprise/TerminologyBiblioTech/CATALOGUE/UMLS_MT_MRCONSO/GLOSSARY_UMLS_MT_MRCONSO_QUERY_TERMS.csv")
}
\description{
This function will make sure that the GLOSSARY and list of files in NETS/ match.
}
