% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_mrconso_net_glossary_2.R
\name{qa_mrconso_net_glossary_2}
\alias{qa_mrconso_net_glossary_2}
\title{This function will deduplicate the GLOSSARY based on the UMLS_QUERY_KEYWORD only}
\usage{

  qa_mrconso_net_glossary_2(path_to_glossary = "/Users/meerapatel/GitHub/MSK_KMI_Enterprise/TerminologyBiblioTech/CATALOGUE/UMLS_MT_MRCONSO/GLOSSARY_UMLS_MT_MRCONSO_QUERY_TERMS.csv")
}
\description{
This function will deduplicate the GLOSSARY based on the UMLS_QUERY_KEYWORD only
}
