% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_mrconso_data_exact_limit_one.R
\name{query_mrconso_data_exact_limit_one}
\alias{query_mrconso_data_exact_limit_one}
\title{Queries MRCONSO UMLS table for the exact phrase and limit 1}
\usage{
query_mrconso_data_exact_limit_one(search_word)
}
\description{
Queries MRCONSO UMLS table for the exact phrase and limit 1
}
