% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_record_from_glossary.R
\name{remove_record_from_glossary}
\alias{remove_record_from_glossary}
\title{Completely removes all traces of a query search}
\usage{
remove_record_from_glossary(umls_sql_keyword_id = NULL, umls_sql_keyword,
  path_to_glossary = "/Users/meerapatel/GitHub/MSK_KMI_Enterprise/TerminologyBiblioTech/CATALOGUE/UMLS_MT_MRCONSO/GLOSSARY_UMLS_MT_MRCONSO_QUERY_TERMS.csv")
}
\arguments{
\item{sql_query_keyword_id}{id number that identifies the keyword. If NULL, all records with parameter sql_query_keyword will be filtered out instead.}

\item{sql_query_keyword}{keyword to remove}
}
\value{
An updated glossary that is written at its existing file destination
}
\description{
Completely removes all traces of a query search
}
