% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_source_concept_from_key.R
\name{remove_source_concept_from_key}
\alias{remove_source_concept_from_key}
\title{Completely removes all traces of a concept that didn't need to be mapped from the key}
\usage{
remove_source_concept_from_key(identity_id, key_field, key_concept_name)
}
\arguments{
\item{identity_id}{identity_id to which the concept belongs}

\item{key_field}{the field this concept belongs to}

\item{key_concept_name}{the exact name of the concept}
}
\value{
An updated KEY that is written at its existing file destination that has removed all record of the given source concept
}
\description{
Completely removes all traces of a concept that didn't need to be mapped from the key
}
