% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Creates a REPORT that contains TABLE_OF_CONTENTS sourced from catalogue, the RAW_IDENTITY, and the MAPPED, PARKED_FOR_DENOVO, and PARKED_FOR_EXTERNAL, and QUEUE statuses}
\usage{
report(path_to_identity_csv, path_to_key_csv,
  filter_out_blank_concepts = TRUE)
}
\description{
Creates a REPORT that contains TABLE_OF_CONTENTS sourced from catalogue, the RAW_IDENTITY, and the MAPPED, PARKED_FOR_DENOVO, and PARKED_FOR_EXTERNAL, and QUEUE statuses
}
