% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_coordinates.R
\name{review_coordinates}
\alias{review_coordinates}
\title{Makes sure that the coordinate cui and its associated strings match up to the keyword. This applies especially to words like "No" that can mean either "Nitrous Oxide" or "No".}
\usage{
review_coordinates(umls_sql_keyword,
  path_to_glossary = "/Users/meerapatel/GitHub/MSK_KMI_Enterprise/TerminologyBiblioTech/CATALOGUE/UMLS_MT_MRCONSO/GLOSSARY_UMLS_MT_MRCONSO_QUERY_TERMS.csv")
}
\arguments{
\item{sql_query_keyword_id}{id number that identifies the keyword.}

\item{sql_query_keyword}{keyword to review}
}
\value{
the UMLS MRCONSO STR's that are associated with the coordinate cui
}
\description{
Makes sure that the coordinate cui and its associated strings match up to the keyword. This applies especially to words like "No" that can mean either "Nitrous Oxide" or "No".
}
