% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographr.R
\name{save_map}
\alias{save_map}
\title{Save map}
\usage{
save_map(plot, filename)
}
\arguments{
\item{plot}{ggplot object}

\item{filename}{filename}

\item{orientation}{'portrait' or 'landscape'}
}
\description{
Save map on disk
}
