% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scavenge_na_concepts.R
\name{scavenge_na_concepts}
\alias{scavenge_na_concepts}
\title{Scavenges and assigns all NA concepts to the "C0027366" cui}
\usage{
scavenge_na_concepts(progress_list, na_values = c("not available", "N/A",
  "n/a", "na"))
}
\arguments{
\item{na_values}{character vector of source values considered to be NA}
}
\description{
Scavenges and assigns all NA concepts to the "C0027366" cui
}
