% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_acknowledgments}
\alias{set_acknowledgments}
\title{Set or Get Acknowledgments Setting}
\usage{
set_acknowledgments(ack = NULL)
}
\arguments{
\item{ack}{A logical value to set the acknowledgments setting. If \code{NULL}, the current setting is returned. The default is \code{NULL}.}
}
\value{
If \code{ack} is \code{NULL}, returns the current acknowledgments setting. If \code{ack} is a logical value, the function will set the acknowledgments setting to that value and return invisibly.
}
\description{
This function sets a new value for the acknowledgments setting in the cartographr environment or retrieves the current setting if no argument is provided.
}
\examples{
\dontrun{
# To get the current acknowledgments setting
set_acknowledgments()

# To set the acknowledgments setting to TRUE
set_acknowledgments(TRUE)
}

}
