% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_attribution}
\alias{set_attribution}
\title{Set or Get Attribution Setting}
\usage{
set_attribution(attribution = NULL)
}
\arguments{
\item{attribution}{A logical value to set the acknowledgments setting. If \code{NULL}, the current setting is returned. The default is \code{NULL}.}
}
\value{
If \code{attribution} is \code{NULL}, returns the current acknowledgments setting. If \code{attribution} is a logical value, the function will set the acknowledgments setting to that value and return invisibly.
}
\description{
This function sets a new value for the attribution setting in the cartographr environment or retrieves the current setting if no argument is provided.
}
\examples{
\dontrun{
# To get the current acknowledgments setting
set_attribution()

# To set the acknowledgments setting to TRUE
set_attribution(TRUE)
}

}
