% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_key_location.R
\name{set_key_location}
\alias{set_key_location}
\title{Sets the key location to iterate on when making updates}
\usage{
set_key_location(path, is.gsheet = TRUE, file_path = !is.gsheet)
}
\arguments{
\item{path}{path to either googlesheet in the format googlesheet_name.tabname or full file path on local}

\item{is.gsheet}{boolean. TRUE if KEY is a Google Sheet.}
}
\description{
Sets the key location to iterate on when making updates
}
