% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_paths.R
\name{set_paths}
\alias{set_paths}
\title{Sets the key location to iterate on when making updates}
\usage{
set_paths(new_catalogue_path = NULL, new_key_path = NULL)
}
\arguments{
\item{path}{path to either googlesheet in the format googlesheet_name.tabname or full file path on local}

\item{is.gsheet}{boolean. TRUE if KEY is a Google Sheet.}
}
\description{
Sets the key location to iterate on when making updates
}
