% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soho_boundary}
\alias{soho_boundary}
\title{SoHo Boundary Simple Features Vector}
\format{
An object of class \code{sf} (inherits from \code{data.frame}),
representing the SoHo boundary with its spatial attributes.
}
\source{
https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-line-file.html
}
\usage{
soho_boundary
}
\description{
This dataset represents the Simple Features vector for the MN24 region of New York,
specifically covering the SoHo neighborhood. It includes spatial boundaries and
other relevant geographical attributes.
}
\examples{
data(soho_boundary)
plot(soho_boundary$geometry)
}
\keyword{datasets}
