% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dataframe_by_identifier.R
\name{split_dataframe_by_identifier}
\alias{split_dataframe_by_identifier}
\title{Splits a dataframe into a list based on an identifier}
\usage{
split_dataframe_by_identifier(dataframe, id_variable)
}
\arguments{
\item{id_variable}{identifier variable to split by}
}
\description{
Splits a dataframe into a list based on an identifier
}
