% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_new_cui_in_key.R
\name{store_new_cui_in_local_key}
\alias{store_new_cui_in_local_key}
\title{Appends local key with new concept annotation}
\usage{
store_new_cui_in_local_key(path_to_key_csv, identity_id, key_field,
  key_concept_name, key_cui = "", key_cui_mrconso_str = "",
  key_cui_justification = "", key_cui_flag_for_review = "",
  key_denovo_cui_justification = "",
  key_denovo_cui_suggested_definition = "",
  key_denovo_cui_assigned = "", key_external_followup_flag = "",
  key_external_followup_request = "",
  key_external_followup_question = "",
  key_external_followup_timestamp = "")
}
\description{
Appends local key with new concept annotation
}
