% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_match_with_rearrangement.R
\name{str_match_with_rearrangement}
\alias{str_match_with_rearrangement}
\title{Inverts concepts with a single dash to see if a string match in MRCONSO can occur with a rearrangement, typically with drug combinations with {asdfasd}-{asdfadsf} naming convention}
\usage{
str_match_with_rearrangement(progress_list)
}
\description{
Inverts concepts with a single dash to see if a string match in MRCONSO can occur with a rearrangement, typically with drug combinations with {asdfasd}-{asdfadsf} naming convention
}
