% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcards_utils.R
\name{theme_postcard}
\alias{theme_postcard}
\title{Postcard theme}
\usage{
theme_postcard(scaling = get_scaling("A4"))
}
\arguments{
\item{scale_factor}{The scale factor for output}
}
\value{
The theme
}
\description{
Description of the function
}
