% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_poster_anton}
\alias{theme_poster_anton}
\title{Apply poster theme with Anton font}
\usage{
theme_poster_anton()
}
\value{
A 'ggplot2' theme object with the poster theme settings applied and 'Anton' font set.
}
\description{
This function applies the poster theme using the 'Anton' font.
}
\examples{
data("osm")
my_map <- osm |> plot_map() +
 theme_poster_anton()

}
