% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_key_with_new_source_values.R
\name{update_key_with_new_source_values}
\alias{update_key_with_new_source_values}
\title{Update a key dataframe with new source values in source dataframe using a common column name}
\usage{
update_key_with_new_source_values(source_dataframe, key_dataframe,
  col_name)
}
\arguments{
\item{col_name}{name of a column common to both dataframes}
}
\description{
Update a key dataframe with new source values in source dataframe using a common column name
}
