% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_progress_list_nrow.R
\name{view_progress_list_nrow}
\alias{view_progress_list_nrow}
\title{View the rownumber per status in progress_list}
\usage{
view_progress_list_nrow(progress_list)
}
\description{
View the rownumber per status in progress_list
}
