% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mrconso_coordinate.R
\name{write_mrconso_coordinate}
\alias{write_mrconso_coordinate}
\title{This function takes a column in a dataframe that contains terms to do an exact string match}
\usage{
write_mrconso_coordinate(dataframe, sql_term_col)
}
\arguments{
\item{dataframe}{dataframe that contains a column of search terms}

\item{sql_term_col}{variable that contains the single string of search term}
}
\description{
This function takes a column in a dataframe that contains terms to do an exact string match
}
