% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mrconso_net.R
\name{write_mrconso_net}
\alias{write_mrconso_net}
\title{Writes a MRCONSO NET file if one does not exist for the given phrase}
\usage{
write_mrconso_net(phrase, rm_forward_slash = TRUE)
}
\description{
Writes a MRCONSO NET file if one does not exist for the given phrase
}
