% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_store_new_cui_in_key.R
\name{bulk_store_new_cui_in_local_key}
\alias{bulk_store_new_cui_in_local_key}
\title{Appends local key with new concept annotation using a dataframe input instead of a single line}
\usage{
bulk_store_new_cui_in_local_key(path_to_key_csv, dataframe, key_cui = "",
  key_cui_mrconso_str = "", key_cui_justification = "",
  key_cui_flag_for_review = "", key_denovo_cui_justification = "",
  key_denovo_cui_suggested_definition = "",
  key_denovo_cui_assigned = "", key_external_followup_flag = "",
  key_external_followup_request = "",
  key_external_followup_question = "",
  key_external_followup_timestamp = "")
}
\arguments{
\item{dataframe}{dataframe with columns IDENTITY_ID, KEY_FIELD, and KEY_CONCEPT_NAME}
}
\description{
Appends local key with new concept annotation using a dataframe input instead of a single line
}
