% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crime}
\alias{crime}
\title{Manhattan crime dataset}
\format{
An object of class \code{data.frame} with 31620 rows and 4 columns.
}
\usage{
crime
}
\description{
This dataset includes all valid felony, misdemeanor, and violation crimes reported to the New York City Police Department (NYPD) for Manhattan from 2004.
\url{https://catalog.data.gov/}.
}
\keyword{datasets}
