% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographr.R
\name{get_osmdata}
\alias{get_osmdata}
\title{Retrieve OSM data for urban environments}
\usage{
get_osmdata(lat, lon, y_distance, x_distance, quiet = F)
}
\arguments{
\item{lat}{Latitude WGS84}

\item{lon}{Longitude WGS84}

\item{y_distance}{Y distance in meters}

\item{x_distance}{X distance in meters}
}
\value{
The shapefiles of buildings, streets, water, beaches, greens, ..
}
\description{
This function retrieves OSM data with sensible defaults for cities
}
