% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographr.R
\name{get_osmdata}
\alias{get_osmdata}
\title{Retrieve OpenStreetMap data for a specified location}
\usage{
get_osmdata(
  lat = NULL,
  lon = NULL,
  x_distance = NULL,
  y_distance = NULL,
  aspect_ratio = NULL,
  bbox = NULL,
  sf = NULL,
  quiet = F
)
}
\arguments{
\item{lat}{Optional numeric, the latitude of the center point for the OSM data retrieval.}

\item{lon}{Optional numeric, the longitude of the center point for the OSM data retrieval.}

\item{x_distance}{Optional numeric, the distance in meters from the center point
to the edge of the area in the x direction (longitude). Default is NULL.}

\item{y_distance}{Optional numeric, the distance in meters from the center point
to the edge of the area in the y direction (latitude). Default is NULL.}

\item{aspect_ratio}{Optional numeric, the ratio of x_distance to y_distance.
Default is NULL.}

\item{bbox}{Optional bbox of the map boundary.
Default is NULL.}

\item{sf}{Optional sf of the map boundary.
Default is NULL.}

\item{quiet}{Logical, whether to suppress informational messages during execution.
Default is FALSE.}
}
\value{
A list containing various elements of the OSM data, including street networks,
buildings, water bodies, green areas, beaches, parking areas, railways, and the
bounding box of the retrieved area.
}
\description{
This function fetches OpenStreetMap (OSM) data for a specified latitude and longitude,
within a given distance range. It allows the user to specify distances in the x and y
directions, an aspect ratio, or a combination of these to define the area for which
OSM data is retrieved. The function ensures that only two of the three parameters
(x_distance, y_distance, aspect_ratio) are provided to define the area.
}
\examples{
\donttest{
osm_data <- get_osmdata(lat = 48.2082, lon = 16.3738, x_distance = 1000)
}
}
