% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcards_utils.R
\name{get_osmdata}
\alias{get_osmdata}
\title{Retrieve OSM data}
\usage{
get_osmdata(center, y_distance, x_distance)
}
\arguments{
\item{center}{Center coordinate}

\item{y_distance}{Y distance in meters}

\item{x_distance}{X distance in meters}
}
\value{
The shapefiles of buildings, streets, ..
}
\description{
This function retrieves OSM data
}
