% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{get_palette}
\alias{get_palette}
\title{Create a Color Theme for Maps}
\usage{
get_palette(palette)
}
\arguments{
\item{palette}{The color palette to use.
Can be one of "alphabet", "arctic", "autumn", "autumn-muted", "bw", "desert", "evening", "gray", "iberia", "imhof", "lines", "metropolitan", "midnight", "minimal", "swiss", "tropical", or a named list for a custom palette.
If a named list is provided, it should contain color hex codes for each map element.
If \code{NULL} or an unrecognized name is provided, the function will throw an error.}
}
\value{
A list containing color settings for the map elements.
}
\description{
This function creates a color theme to be used with \code{plot_map()}. It can accept a predefined palette name or a custom palette provided as a named list.
}
\details{
The color moods for the predefined palettes are described as follows:
\itemize{
\item \strong{Alphabet}: A modern palette with a straightforward aesthetic.
\item \strong{Arctic}: A palette that reflects the clear and bright qualities of Arctic landscapes.
\item \strong{Autumn}: A palette with the warm and varied hues typical of the fall season.
\item \strong{Autumn Muted}: A subdued version of the autumn palette, offering a gentle and warm atmosphere.
\item \strong{BW}: A classic black and white palette with a hint of warmth for a traditional feel.
\item \strong{Desert}: A palette that captures the earthy and natural tones found in desert environments.
\item \strong{Evening}: A palette that embodies the quiet and contemplative nature of dusk.
\item \strong{Gray}: A balanced palette that provides a composed and refined look.
\item \strong{Iberia}: A palette that reflects the warm and diverse tones associated with the Iberian landscape.
\item \strong{Imhof}: A palette with natural and subdued tones, inspired by the work of cartographer Eduard Imhof.
\item \strong{Lines (BW)}: A contrasting black and white palette for a clear and defined appearance.
\item \strong{Metropolitan}: A palette with understated tones that suggest urban sophistication.
\item \strong{Midnight}: A palette that conveys the depth and tranquility of the night.
\item \strong{Minimal}: A palette focused on minimalism, utilizing primarily whites and light grays.
\item \strong{Swiss}: A palette that emphasizes cleanliness and precision, reminiscent of Swiss design.
\item \strong{Tropical}: A lively palette with the bright and bold colors characteristic of tropical areas.
}
}
\examples{
\dontrun{
# Standard usage with a character argument for a predefined palette
plot_map(osm, palette = 'tropical')

# Custom palette creation using a named list for a simple black and white palette
custom_palette <- list(
  palette_building = c("#000000", "#FFFFFF"),
  water = "#000000",
  green = "#FFFFFF",
  beach = "#000000",
  parking = "#FFFFFF",
  street = "#000000",
  background = "#FFFFFF",
  railway = "#000000",
  hatched = TRUE
)
plot_map(osm, palette = custom_palette)
}
}
