% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identity_from_csv.R
\name{identity_from_csv}
\alias{identity_from_csv}
\title{Gets IDENTITY data and cleans up based on parameters}
\usage{
identity_from_csv(path_to_identity, target_cols,
  redcap_forms_to_exclude = NULL, keep_all_cols = FALSE,
  dont_trimws = FALSE, remove_all = "\\t|\\n",
  include_duplicates = FALSE, log = TRUE)
}
\arguments{
\item{path_to_identity}{path to  the rawidentity csv}

\item{target_cols}{vector of the names of the columns that will be melted into the "KEY_FIELD" variable}

\item{keep_all_cols}{TRUE if all the variables are wanted in the output. Otherwise, only the target_cols will be selected.}

\item{dont_trimws}{TRUE if white spaces on both the left and right side are not desired in the output}

\item{remove_all}{character string of length 1 representing the string of artifact and other values to be removed from the final output separated by the OR operator}

\item{include_duplicates}{TRUE if all the duplicates are desired in the output}
}
\value{
dataframe where raw REDCap data dictionary dataframe is processed to a concept level where the target_cols parameter identifies the target columns in the raw data as concepts
}
\description{
Gets IDENTITY data and cleans up based on parameters
}
