% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_from_csv.R, R/key_from_xlsx.R
\name{key_from_csv}
\alias{key_from_csv}
\title{Gets KEY data and cleans up based on parameters}
\usage{
key_from_csv(path_to_key, tab = 1, log = TRUE)

key_from_csv(path_to_key, tab = 1, log = TRUE)
}
\arguments{
\item{path_to_key}{full file path to key csv file}

\item{log}{TRUE if this event will be logged using the log functions from the mirCat package}

\item{path_to_key}{full file path to key csv file}

\item{log}{TRUE if mirCat::log functions will be used to log this event}
}
\description{
Gets KEY data and cleans up based on parameters

Gets KEY data and cleans up based on parameters
}
