% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcards_utils.R
\name{plot_map}
\alias{plot_map}
\title{Plot a map}
\usage{
plot_map(
  osm,
  theme = get_theme("imhof", "Poppins"),
  scaling = get_scaling("A4"),
  circle = FALSE
)
}
\arguments{
\item{osm}{OSM object to plot}

\item{scaling}{scaling applied to the plot}

\item{circle}{If TRUE draw circle plot}

\item{color}{Color theme applied to the plot}
}
\description{
This function plots a map
}
