% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographr.R
\name{plot_map}
\alias{plot_map}
\title{Plot a map}
\usage{
plot_map(
  osm,
  color = get_color("imhof", "Poppins"),
  scaling = get_scaling("A4")
)
}
\arguments{
\item{osm}{OSM object to plot}

\item{color}{Color theme applied to the plot}

\item{scaling}{scaling applied to the plot}

\item{circle}{If TRUE draw circle plot}
}
\description{
This function plots a map
}
