% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographr.R
\name{plot_map}
\alias{plot_map}
\title{Wrapper function for plot_map}
\usage{
plot_map(osm, palette = "imhof")
}
\arguments{
\item{osm}{OSM object to plot}

\item{palette}{Color theme applied to the plot}
}
\description{
This function plots a map
}
