% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcards_utils.R
\name{plot_map}
\alias{plot_map}
\title{Plot a map}
\usage{
plot_map(osm_object, color, scaling)
}
\arguments{
\item{osm_object}{OSM object to plot}

\item{color}{Color theme applied to the plot}

\item{scaling}{scaling applied to the plot}
}
\description{
This function plots a map
}
