% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographr.R
\name{preprocess_map}
\alias{preprocess_map}
\title{Preprocess OpenStreetMap data}
\usage{
preprocess_map(osm)
}
\arguments{
\item{osm}{A list object containing OSM data elements to be preprocessed.}
}
\value{
The preprocessed OSM object with validated and unified water bodies and buildings,
and with additional preprocessing information.
}
\description{
This function performs preprocessing on OpenStreetMap (OSM) data. It validates and
unifies various OSM elements such as water lines, polygons, and multipolygons, as well
as building polygons and multipolygons.
}
\examples{
data("osm")
preprocessed_osm <- preprocess_map(osm)

}
