% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcards_utils.R
\name{save_map}
\alias{save_map}
\title{Save map}
\usage{
save_map(p, outdir, osm_object, scaling, color, format)
}
\arguments{
\item{p}{ggplot object}

\item{outdir}{output_filename}

\item{osm_object}{osm_object}

\item{scaling}{scaling}

\item{color}{color}

\item{format}{format}
}
\description{
Save map on disk
}
