% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_infomap_poppins}
\alias{theme_infomap_poppins}
\title{Apply infomap Theme with Poppins Font}
\usage{
theme_infomap_poppins()
}
\value{
A 'ggplot2' theme object with the infomap theme settings applied and 'Poppins' font set.
}
\description{
This function applies the infomap theme using the 'Poppins' font.
}
\examples{
\dontrun{
# Apply the infomap theme with Poppins font to a ggplot object
ggplot_object + theme_poster_poppins()
}
}
