% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartographr.R
\name{theme_infomap_poppins}
\alias{theme_infomap_poppins}
\title{Infomap theme with font Anton}
\usage{
theme_infomap_poppins()
}
\arguments{
\item{scale_factor}{The scale factor for output}
}
\value{
The theme
}
\description{
Description of the function
}
