% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_poster_poppins}
\alias{theme_poster_poppins}
\title{Apply poster theme with Poppins font}
\usage{
theme_poster_poppins()
}
\value{
A 'ggplot2' theme object with the poster theme settings applied and 'Poppins' font set.
}
\description{
This function applies the poster theme using the 'Poppins' font.
}
\examples{
data("osm")
my_map <- osm |> plot_map() +
 theme_poster_poppins()

}
